add shared project

+++ test/HttpProblemDetails.Nancy.Tests/InsufficientCashException.cs
@@ -5,7 +5,7 @@ namespace HttpProblemDetails.Nancy.Tests
     public class InsufficientCashException : HttpProblemDetailException<InsufficientCashProblem>
     {
         public InsufficientCashException(HttpStatusCode statusCode, IHttpProblemDetail problemDetail) 
-            : base(statusCode, problemDetail)
+            : base((int)statusCode, problemDetail)
         { }
     }
 }
